+++ data-raw/prepare_aphid.R
@@ -1,5 +1,6 @@
-# make example data: simulate individual data from pooled development days data in
-# Satar, S., and Yokomi, R., 2002.Effect of Temperature and Host on Development of Brachycaudus schwartzi (Homoptera: Aphididae).
+# make example data: simulate individual data from pooled development days data
+# in Satar, S., and Yokomi, R., 2002.Effect of Temperature and Host on
+# Development of Brachycaudus schwartzi (Homoptera: Aphididae).
 # Ann. Entomol. Soc. Am. 95(5): 597-602
 #