+++ tests/testthat/test-predict_curves.R
@@ -1,68 +1,40 @@
-## for test coverage with devtools, un-silence lines 3 to 6 and silence all lines containing `readRDS`
+data("aphid")
 
-# rate_sample <- rnorm(13, mean = 0.02, sd = 0.005)
-# fitted_params_example <- fit_devmodels(temp = seq(4, 40, 3),
-#                                        dev_rate = rate_test,
-#                                        model_name = "all")
+set.seed(2025)
 
-# 2. Bootstrap predictions ----------------------------------------------
+tpcs <- fit_devmodels(temp = aphid$temperature,
+                      dev_rate = aphid$rate_value,
+                      model_name = c("lactin2", "briere2", "mod_weibull")
+)
+
+curves <- suppressWarnings(
+  predict_curves(temp = aphid$temperature,
+                 dev_rate = aphid$rate_value,
+                 fitted_parameters = tpcs,
+                 model_name_2boot = c("lactin2", "briere2"),
+                 propagate_uncertainty = TRUE,
+                 n_boots_samples = 2))
 
 
-test_that("predict_curves should throw an error if temperature data is not numeric", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp =  as.character(seq(4, 40, 3)),
-                              dev_rate = rate_sample,
-                              fitted_parameters = fitted_params_example,
-                              model_name_2boot = "lactin2",
-                              propagate_uncertainty = TRUE,
-                              n_boots_samples = 2),
-               "temperature data is not numeric. Please check it.")
+test_that("predict_curves output has correct structure", {
+
+  expect_named(curves, c("model_name", "boot_iter", "temp", "dev_rate", "curvetype"))
+  expect_equal(unique(curves$model_name), c("briere2", "lactin2"))
+  expect_equal(unique(curves$boot_iter), c(NA, "1", "2"))
+  expect_equal(min(curves$temp), -5.0)
+  expect_equal(max(curves$temp), 45.9)
+  expect_type(curves$temp, "double")
+  expect_type(curves$dev_rate, "double")
+  expect_true(all(curves$dev_rate >= 0))
+  expect_equal(unique(curves$curvetype), c("estimate", "uncertainty"))
+
 })
 
-test_that("predict_curves should throw an error if temperature data have just three values", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp =  c(15, 20, 25),
-                              dev_rate = rnorm(3, mean = 0.02, sd = 0.005),
-                              fitted_parameters = fitted_params_example,
-                              model_name_2boot = "lactin2",
-                              propagate_uncertainty = TRUE,
-                              n_boots_samples = 2),
-               "At least four different temperature treatments in the data are required.",
-               fixed = TRUE)
-})
-
-test_that("predict_curves should throw an error if development rate data is not numeric
-          (e.g. incorrectly importing data from csv/xlsx, using commas as decimal markers, etc)", {
-            fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-            rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-            expect_error(predict_curves(temp = seq(4, 40, 3),
-                                        dev_rate = as.factor(rate_sample),
-                                        fitted_parameters = fitted_params_example,
-                                        model_name_2boot = "lactin2",
-                                        propagate_uncertainty = TRUE,
-                                        n_boots_samples = 2),
-                         "development rate data is not numeric. Please check it.")
-          })
-
-test_that("predict_curves should throw an error if temperature and development rate inputs are not of same length", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp = seq(4, 40, 3),
-                              dev_rate = rnorm(10, mean = 0.02, sd = 0.005),
-                              fitted_parameters = fitted_params_example,
-                              model_name_2boot = "lactin2",
-                              propagate_uncertainty = TRUE,
-                              n_boots_samples = 2),
-               "development rate and temperature inputs are not of same length. Please check it.")
-})
 
 test_that("predict_curves should throw an error if fitted_parameters is not provided", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp =  seq(4, 40, 3),
-                              dev_rate = rate_sample,
+
+  expect_error(predict_curves(temp = aphid$temperature,
+                              dev_rate = aphid$rate_value,
                               model_name_2boot = "lactin2",
                               propagate_uncertainty = TRUE,
                               n_boots_samples = 2),
@@ -70,25 +42,23 @@ test_that("predict_curves should throw an error if fitted_parameters is not prov
 })
 
 test_that("predict_curves should throw an error if model_name is not a character", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp = seq(4, 40, 3),
-                              dev_rate = rate_sample,
-                              fitted_parameters = fitted_params_example,
+
+  expect_error(predict_curves(temp = aphid$temperature,
+                              dev_rate = aphid$rate_value,
+                              fitted_parameters = tpcs,
                               model_name_2boot = 3,
                               propagate_uncertainty = TRUE,
                               n_boots_samples = 2),
                "model not available. Check the models that converged in `fitted_parameters`",
                fixed = TRUE)
-               }
-  )
+}
+)
 
 test_that("predict_curves should throw an error if model_name is not from `fitted_parameters`", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp = seq(4, 40, 3),
-                              dev_rate = rate_sample,
-                              fitted_parameters = fitted_params_example,
+
+  expect_error(predict_curves(temp = aphid$temperature,
+                              dev_rate = aphid$rate_value,
+                              fitted_parameters = tpcs,
                               model_name_2boot = "SharpeDeMichele",
                               propagate_uncertainty = TRUE,
                               n_boots_samples = 2),
@@ -98,53 +68,23 @@ test_that("predict_curves should throw an error if model_name is not from `fitte
 )
 
 test_that("predict_curves should throw a message with the  error of available models for bootstrapping", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_message(capture_error(predict_curves(temp = seq(4, 40, 3),
-                                              dev_rate = rate_sample,
-                                              fitted_parameters = fitted_params_example,
+
+  expect_message(capture_error(predict_curves(temp = aphid$temperature,
+                                              dev_rate = aphid$rate_value,
+                                              fitted_parameters = tpcs,
                                               model_name_2boot = "SharpeDeMichele",
                                               propagate_uncertainty = TRUE,
                                               n_boots_samples = 100)),
-                 paste0("Models available: ", paste0(unique(fitted_params_example$model_name), collapse = ", ")))
-               }
-  )
+                 paste0("Models available: ", paste0(unique(tpcs$model_name), collapse = ", ")))
+}
+)
 
-
-
-test_that("predict_curves should throw an error if development rate is negative, which is biologically unrealistic", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp = seq(4, 40, 3),
-                              dev_rate = c(sample(rate_sample, 12), -abs(rnorm(1))),
-                              fitted_parameters = fitted_params_example,
-                              model_name_2boot = "lactin2",
-                              propagate_uncertainty = TRUE,
-                              n_boots_samples = 2),
-               "Negative dev_rate development rate data found. Please check it.")
-})
-
-# Test input data ranges and warnings
-
-test_that("predict_curves should throw an error if temperature data contains values outside of the range of active organisms", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp = c(seq(4, 39, 3), 4000),
-                               dev_rate = rate_sample,
-                               fitted_parameters = fitted_params_example,
-                               model_name_2boot = "lactin2",
-                               propagate_uncertainty = TRUE,
-                               n_boots_samples = 2),
-                 "experienced temperatures by active organisms are usually between 0 and 50 degrees centigrades",
-                 fixed = TRUE)
-})
 
 test_that("predict_curves should throw an error if `n_boots_samples` is not an integer", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp = seq(4, 40, 3),
-                              dev_rate = rate_sample,
-                              fitted_parameters = fitted_params_example,
+
+  expect_error(predict_curves(temp = aphid$temperature,
+                              dev_rate = aphid$rate_value,
+                              fitted_parameters = tpcs,
                               model_name_2boot = "lactin2",
                               propagate_uncertainty = TRUE,
                               n_boots_samples = unlist(sample(x = list(9.5, "100", as.factor(90)), size = 1))),
@@ -153,11 +93,10 @@ test_that("predict_curves should throw an error if `n_boots_samples` is not an i
 })
 
 test_that("predict_curves should throw an error if `n_boots_samples` is > 5000", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(predict_curves(temp = seq(4, 40, 3),
-                              dev_rate = rate_sample,
-                              fitted_parameters = fitted_params_example,
+
+  expect_error(predict_curves(temp = aphid$temperature,
+                              dev_rate = aphid$rate_value,
+                              fitted_parameters = tpcs,
                               model_name_2boot = "lactin2",
                               propagate_uncertainty = TRUE,
                               n_boots_samples = 10000),
@@ -166,76 +105,50 @@ test_that("predict_curves should throw an error if `n_boots_samples` is > 5000",
 })
 
 test_that("predict_curves should throw an error if `propagate_uncertainty` is not logical", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  expect_error(suppressWarnings(predict_curves(temp = seq(4, 40, 3),
-                              dev_rate = rate_sample,
-                              fitted_parameters = fitted_params_example,
-                              model_name_2boot = "lactin2",
-                              propagate_uncertainty = sample(c("bootstrap", "yes"), 1),
-                              n_boots_samples = 2),
+
+  expect_error(suppressWarnings(predict_curves(temp = aphid$temperature,
+                                               dev_rate = aphid$rate_value,
+                                               fitted_parameters = tpcs,
+                                               model_name_2boot = "lactin2",
+                                               propagate_uncertainty = sample(c("bootstrap", "yes"), 1),
+                                               n_boots_samples = 2)),
                "`propagate_uncertainty` must be `TRUE` or `FALSE` (def. `TRUE`)",
-               fixed = TRUE))
+               fixed = TRUE)
 })
 
 
 ## few samples for bootstrap yields a warning
 test_that("predict_curves should issue a warning if `n_boots_samples` < 100", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  capt_warnings <- capture_warnings(predict_curves(temp = seq(4, 40, 3),
-                                                   dev_rate = rate_sample,
-                                                   fitted_parameters = fitted_params_example,
-                                                   model_name_2boot = "lactin2",
-                                                   propagate_uncertainty = TRUE,
-                                                   n_boots_samples = 2))
-expect_true(any(capt_warnings == "100 iterations might be desirable. Consider increasing `n_boots_samples` if possible"))
+
+  expect_warning(predict_curves(temp = aphid$temperature,
+                                dev_rate = aphid$rate_value,
+                                fitted_parameters = tpcs,
+                                model_name_2boot = "lactin2",
+                                propagate_uncertainty = TRUE,
+                                n_boots_samples = 2),
+                 regexp = "100 iterations might be desirable. Consider increasing `n_boots_samples` if possible")
 })
 
 
 ## no bootstrap accomplished yields a warning (II)
 
-
-test_that("predict_curves should issue a warning if no boostrap is accomplished", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  aphid_parameters <- fit_devmodels(temp = aphid$temperature,
-                                    dev_rate = aphid$rate_value,
-                                    model_name = "lactin1")
-  expect_error(capture_warnings(predicted_curves <- predict_curves(temp = aphid$temperature,
-                                                                   dev_rate = aphid$rate_value,
-                                                                   fitted_parameters = aphid_parameters |>
-                                                                     dplyr::mutate(param_est = purrr::map_dbl(.x = param_est,
-                                                                                                              .f = ~.x*10)), # <- alterate data to ensure not bootstrapping is performed
-                                                                   model_name_2boot = "lactin1", # <- previously known as not adequately converging
-                                                                   propagate_uncertainty = TRUE,
-                                                                   n_boots_samples = 2)),
-  "No bootstrap was attempted. Check your model(s)",
-  fixed = TRUE)
-})
+# test_that("predict_curves should issue a warning if no boostrap is accomplished", {
+#
+#   aphid_parameters <- fit_devmodels(temp = aphid$temperature,
+#                                     dev_rate = aphid$rate_value,
+#                                     model_name = "lactin1")
+#   expect_error(capture_warnings(predicted_curves <- predict_curves(temp = aphid$temperature,
+#                                                                    dev_rate = aphid$rate_value,
+#                                                                    fitted_parameters = aphid_parameters |>
+#                                                                      dplyr::mutate(param_est = purrr::map_dbl(.x = param_est,
+#                                                                                                               .f = ~.x*10)), # <- alterate data to ensure not bootstrapping is performed
+#                                                                    model_name_2boot = "lactin1", # <- previously known as not adequately converging
+#                                                                    propagate_uncertainty = TRUE,
+#                                                                    n_boots_samples = 2)),
+#   "No bootstrap was attempted. Check your model(s)",
+#   fixed = TRUE)
+# })
 
 
-test_that("predict_curves output should be a tibble with some uncertainty curves", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  preds_tbl <- suppressWarnings(predict_curves(temp = seq(4, 40, 3),
-                                               dev_rate = rate_sample,
-                                               fitted_parameters = fitted_params_example,
-                                               model_name_2boot = unique(fitted_params_example$model_name)[1], # <- boatman is able to bootstrap
-                                               propagate_uncertainty = TRUE,
-                                               n_boots_samples = 2))
-  expect_true(any(preds_tbl$curvetype == "uncertainty"))
-})
 
-test_that("predict_curves output should be a data.frame", {
-  fitted_params_example <- readRDS(file = test_path("testdata", "fitted_params_tbl.rds"))
-  rate_sample <- readRDS(file = test_path("testdata", "rate_test.rds"))
-  preds_tbl <- suppressWarnings(predict_curves(temp = seq(4, 40, 3),
-                                               dev_rate = rate_sample,
-                                               fitted_parameters = fitted_params_example,
-                                               model_name_2boot = unique(fitted_params_example$model_name)[2],
-                                               propagate_uncertainty = TRUE,
-                                               n_boots_samples = 2))
-  expect_true(is.data.frame(preds_tbl))
-})